producer/requirements.txt
google-cloud-pubsub==2.22.0
